refactor(tiling): remove design mode
The pane-grid editor, its handles, presets, empty slots, undo stack and
spring keys go; titlebar drag-to-slot and edge-drag resize stay, with
the split arithmetic they share moved to tiling/splits.rs.
