+++ src/winit.rs
@@ -828,7 +828,6 @@ pub fn run_winit() {
             // Pick up any tiling tree a close, minimize or workspace move
             // left dirty; a no-op flag read when nothing changed.
             state.flush_tiling_relayout();
-            state.flush_tiling_design();
             state.popups.cleanup();
             // Tell any window that has moved where it is now. Diffed against
             // what was last sent, so a desktop at rest sends nothing.
