+++ src/x11.rs
@@ -566,7 +566,6 @@ pub fn run_x11() {
         } else {
             state.workspaces.refresh_space();
             state.flush_tiling_relayout();
-            state.flush_tiling_design();
             state.popups.cleanup();
             display_handle.flush_clients().unwrap();
         }
